- using size_t for regular index and size variables
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@190 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/prim/point_2d.inl
@@ -95,7 +95,7 @@ Point2D<T>::position() const
 
 template<typename T> inline
 typename Point2D<T>::TConstReference
-Point2D<T>::operator[](unsigned iIndex) const
+Point2D<T>::operator[](size_t iIndex) const
 {
 	LASS_ASSERT(iIndex < dimension);
 	return *(&x + iIndex);
@@ -105,7 +105,7 @@ Point2D<T>::operator[](unsigned iIndex) const
 
 template<typename T> inline
 typename Point2D<T>::TReference
-Point2D<T>::operator[](unsigned iIndex)
+Point2D<T>::operator[](size_t iIndex)
 {
 	LASS_ASSERT(iIndex < dimension);
 	return *(&x + iIndex);
