+++ src/prim/point_3dh.h
@@ -84,8 +84,8 @@ public:
 	explicit Point3DH(const TVector& iPositionVector);
 
 	const TVector position() const;
-	TConstReference operator[](unsigned iIndex) const;
-	TReference operator[](unsigned iIndex);
+	TConstReference operator[](size_t iIndex) const;
+	TReference operator[](size_t iIndex);
 	TConstReference at(signed iIndex) const;
 	TReference at(signed iIndex);
